[escher] Stencil shadow volumes.
This mega CL is a major overhaul of the Escher renderer.
It introduces new classes such as PaperDrawCallFactory,
and refines the relationships between existing classes,
such as PaperRenderer2, PaperShapeCache, and PaperRenderQueue.

PaperShapeCache can now generate specially-constructed
meshes that bake in the ability to extrude shadow volume
geometry, in a way that is independent of the relative
position of the light (and which can therefore support
multiple lights).

ES-133 #done
ES-141 #comment
ES-144 #done
ES-145 #done

Change-Id: Ic08e7e1770712fe3b71e111b919ed759152f9fbd