[scenic] Saner session shutdown.
The goal is to provide a general framework for session-shutdown
requests from within the system (as opposed to shutdown because
the client closed the channel).  The approach is for shutdown
requests to go up to the Scenic session which then calls
PrepareForShutdown() on all the session's CommandDispatchers.
In order to not destroy the session while the call-stack
still references it, a task is posted to destroy the session
asynchronously.

Change-Id: Ib9172c8c6fc6891c9e13ec6a17d7964dfa78f934